+++ packages/grafana-flamegraph/src/CallTree/FlameGraphCallTreeContainer.tsx
@@ -1,36 +1,20 @@
-import { css, cx } from '@emotion/css';
 import { memo, useMemo, useState, useRef, useEffect, useCallback } from 'react';
-import {
-  useTable,
-  useSortBy,
-  useExpanded,
-  Column,
-  Row,
-  UseExpandedRowProps,
-  HeaderGroup,
-  TablePropGetter,
-  TableBodyPropGetter,
-  TableProps,
-  TableBodyProps,
-} from 'react-table';
+import { useTable, useSortBy, useExpanded, Column, Row, UseExpandedRowProps } from 'react-table';
 import AutoSizer from 'react-virtualized-auto-sizer';
 
-import { GrafanaTheme2 } from '@grafana/data';
-import { Button, Dropdown, Icon, IconButton, Menu, Tooltip, useStyles2, useTheme2 } from '@grafana/ui';
+import { Button, Icon, IconButton, Tooltip, useStyles2, useTheme2 } from '@grafana/ui';
 
 import { GetExtraContextMenuButtonsFunction } from '../FlameGraph/FlameGraphContextMenu';
-import { FlameGraphDataContainer, LevelItem } from '../FlameGraph/dataTransform';
+import { FlameGraphDataContainer } from '../FlameGraph/dataTransform';
 import { ColorScheme, ColorSchemeDiff, PaneView, ViewMode } from '../types';
 
-import {
-  buildAllCallTreeNodes,
-  buildCallersTree,
-  CallTreeNode,
-  getInitialExpandedState,
-  getRowBarColor,
-} from './utils';
-
-type Styles = ReturnType<typeof getStyles>;
+import { ActionsCell } from './ActionsCell';
+import { CallTreeTable } from './CallTreeTable';
+import { ColorBarCell } from './ColorBarCell';
+import { DiffCell } from './DiffCell';
+import { FunctionCellWithExpander } from './FunctionCellWithExpander';
+import { getStyles } from './styles';
+import { buildAllCallTreeNodes, buildCallersTree, CallTreeNode, getInitialExpandedState } from './utils';
 
 type Props = {
   data: FlameGraphDataContainer;
@@ -741,754 +725,5 @@ const FlameGraphCallTreeContainer = memo(
 );
 
 FlameGraphCallTreeContainer.displayName = 'FlameGraphCallTreeContainer';
-
-type CallTreeTableProps = {
-  width: number;
-  height: number;
-  compactModeThreshold: number;
-  isCompact: boolean;
-  setIsCompact: (compact: boolean) => void;
-  getFunctionColumnWidth: (availableWidth: number, compactMode: boolean) => number | undefined;
-  getTableProps: (propGetter?: TablePropGetter<CallTreeNode>) => TableProps;
-  getTableBodyProps: (propGetter?: TableBodyPropGetter<CallTreeNode>) => TableBodyProps;
-  headerGroups: Array<HeaderGroup<CallTreeNode>>;
-  rows: Array<Row<CallTreeNode>>;
-  prepareRow: (row: Row<CallTreeNode>) => void;
-  styles: Styles;
-  currentSearchMatchId?: string;
-  searchMatchRowRef: (node: HTMLTableRowElement | null) => void;
-  scrollContainerRef: { current: HTMLDivElement | null };
-  focusedNodeId?: string;
-  callersNodeLabel?: string;
-};
-
-function CallTreeTable({
-  width,
-  height,
-  compactModeThreshold,
-  isCompact,
-  setIsCompact,
-  getFunctionColumnWidth,
-  getTableProps,
-  getTableBodyProps,
-  headerGroups,
-  rows,
-  prepareRow,
-  styles,
-  currentSearchMatchId,
-  searchMatchRowRef,
-  scrollContainerRef,
-  focusedNodeId,
-  callersNodeLabel,
-}: CallTreeTableProps) {
-  const SCROLLBAR_WIDTH = 16;
-  const availableWidth = width - SCROLLBAR_WIDTH;
-  const shouldBeCompact = availableWidth > 0 && availableWidth < compactModeThreshold;
-
-  useEffect(() => {
-    if (availableWidth <= 0) {
-      return;
-    }
-    if (shouldBeCompact !== isCompact) {
-      setIsCompact(shouldBeCompact);
-    }
-  }, [availableWidth, shouldBeCompact, isCompact, setIsCompact]);
-
-  const functionColumnWidth = getFunctionColumnWidth(availableWidth, isCompact);
-
-  if (width < 3 || height < 3) {
-    return null;
-  }
-
-  return (
-    <div style={{ width, height, display: 'flex', flexDirection: 'column' }}>
-      <table {...getTableProps()} className={styles.table} style={{ flexShrink: 0 }}>
-        <thead className={styles.thead}>
-          {headerGroups.map((headerGroup) => {
-            const { key, ...headerGroupProps } = headerGroup.getHeaderGroupProps();
-            return (
-              <tr key={key} {...headerGroupProps}>
-                {headerGroup.headers.map((column) => {
-                  const { key: headerKey, ...headerProps } = column.getHeaderProps(column.getSortByToggleProps());
-                  const columnWidth = column.id === 'label' ? functionColumnWidth : column.width;
-                  return (
-                    <th
-                      key={headerKey}
-                      {...headerProps}
-                      className={styles.th}
-                      style={{
-                        ...(columnWidth !== undefined && { width: columnWidth }),
-                        textAlign: column.id === 'self' || column.id === 'total' ? 'right' : undefined,
-                        ...(column.minWidth !== undefined && { minWidth: column.minWidth }),
-                      }}
-                    >
-                      {column.render('Header')}
-                      <span className={styles.sortIndicator}>
-                        {column.isSorted ? (column.isSortedDesc ? ' ▼' : ' ▲') : ''}
-                      </span>
-                    </th>
-                  );
-                })}
-              </tr>
-            );
-          })}
-        </thead>
-      </table>
-      <div
-        ref={scrollContainerRef}
-        style={{ flex: 1, overflowY: 'scroll', overflowX: 'auto' }}
-        className={styles.scrollContainer}
-      >
-        <table {...getTableProps()} className={styles.table}>
-          <tbody {...getTableBodyProps()} className={styles.tbody}>
-            {rows.map((row, rowIndex) => {
-              prepareRow(row);
-              const { key, ...rowProps } = row.getRowProps();
-              const isFocusedRow = row.original.id === focusedNodeId;
-              const isCallersTargetRow = callersNodeLabel && row.original.label === callersNodeLabel;
-              const isSearchMatchRow = currentSearchMatchId && row.original.id === currentSearchMatchId;
-
-              return (
-                <tr
-                  key={key}
-                  {...rowProps}
-                  ref={isSearchMatchRow ? searchMatchRowRef : null}
-                  className={cx(
-                    styles.tr,
-                    (isFocusedRow ||
-                      (focusedNodeId?.startsWith('label:') && focusedNodeId.substring(6) === row.original.label)) &&
-                      styles.focusedRow,
-                    isCallersTargetRow && styles.callersTargetRow,
-                    isSearchMatchRow && styles.searchMatchRow
-                  )}
-                >
-                  {row.cells.map((cell) => {
-                    const { key: cellKey, ...cellProps } = cell.getCellProps();
-                    const isValueColumn = cell.column.id === 'self' || cell.column.id === 'total';
-                    const isActionsColumn = cell.column.id === 'actions';
-                    const columnWidth = cell.column.id === 'label' ? functionColumnWidth : cell.column.width;
-                    return (
-                      <td
-                        key={cellKey}
-                        {...cellProps}
-                        className={cx(
-                          styles.td,
-                          isActionsColumn && styles.actionsColumnCell,
-                          isValueColumn && styles.valueColumnCell
-                        )}
-                        style={{
-                          ...(columnWidth !== undefined && { width: columnWidth }),
-                          ...(cell.column.minWidth !== undefined && { minWidth: cell.column.minWidth }),
-                        }}
-                      >
-                        {cell.render('Cell', { rowIndex })}
-                      </td>
-                    );
-                  })}
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-}
-
-type ActionsCellProps = {
-  nodeId: string;
-  label: string;
-  itemIndexes: number[];
-  levelItem: LevelItem;
-  hasChildren: boolean;
-  depth: number;
-  parentId: string | undefined;
-  onFocus: (nodeIdOrLabel: string, isLabel: boolean, itemIndexes: number[]) => void;
-  onShowCallers: (label: string) => void;
-  onSearch?: (symbol: string) => void;
-  focusedNodeId: string | undefined;
-  callersNodeLabel: string | undefined;
-  isSearchMatch: boolean;
-  actionsCellClass: string;
-  getExtraContextMenuButtons?: GetExtraContextMenuButtonsFunction;
-  data: FlameGraphDataContainer;
-  viewMode?: ViewMode;
-  paneView?: PaneView;
-  search: string;
-};
-
-const ActionsCell = memo(function ActionsCell({
-  nodeId,
-  label,
-  itemIndexes,
-  levelItem,
-  hasChildren,
-  depth,
-  parentId,
-  onFocus,
-  onShowCallers,
-  onSearch,
-  focusedNodeId,
-  callersNodeLabel,
-  isSearchMatch,
-  actionsCellClass,
-  getExtraContextMenuButtons,
-  data,
-  viewMode,
-  paneView,
-  search,
-}: ActionsCellProps) {
-  const isTheFocusedNode =
-    nodeId === focusedNodeId || (focusedNodeId?.startsWith('label:') && focusedNodeId.substring(6) === label);
-  const isTheCallersTarget = label === callersNodeLabel;
-  const inCallersMode = callersNodeLabel !== undefined;
-  const inFocusMode = focusedNodeId !== undefined;
-  const isRootNode = depth === 0 && !parentId;
-
-  const shouldShowFocusItem = hasChildren && !isTheFocusedNode && !(isRootNode && !inFocusMode);
-  const shouldShowCallersItem = !isTheCallersTarget && !isRootNode;
-  const shouldShowSearchItem = onSearch && !isSearchMatch;
-
-  const extraButtons = useMemo(() => {
-    if (!getExtraContextMenuButtons) {
-      return [];
-    }
-    const clickedItemData = {
-      label,
-      item: levelItem,
-      posX: 0,
-      posY: 0,
-    };
-    return getExtraContextMenuButtons(clickedItemData, data.data, {
-      viewMode: viewMode ?? ViewMode.Single,
-      paneView: paneView ?? PaneView.CallTree,
-      isDiff: data.isDiffFlamegraph(),
-      search,
-    });
-  }, [getExtraContextMenuButtons, label, levelItem, data, viewMode, paneView, search]);
-
-  const hasAnyAction = shouldShowFocusItem || shouldShowCallersItem || shouldShowSearchItem || extraButtons.length > 0;
-
-  if (!hasAnyAction) {
-    return <div className={actionsCellClass} />;
-  }
-
-  const menu = (
-    <Menu>
-      {shouldShowFocusItem && (
-        <Menu.Item
-          label="Focus on callees"
-          icon="compress-arrows"
-          onClick={() => {
-            if (inCallersMode) {
-              onFocus(label, true, itemIndexes);
-            } else {
-              onFocus(nodeId, false, itemIndexes);
-            }
-          }}
-        />
-      )}
-      {shouldShowCallersItem && (
-        <Menu.Item label="Show callers" icon="expand-arrows-alt" onClick={() => onShowCallers(label)} />
-      )}
-      {shouldShowSearchItem && <Menu.Item label="Search" icon="search" onClick={() => onSearch!(label)} />}
-      {extraButtons.map(({ label: btnLabel, icon, onClick }) => (
-        <Menu.Item key={btnLabel} label={btnLabel} icon={icon} onClick={onClick} />
-      ))}
-    </Menu>
-  );
-
-  return (
-    <div className={actionsCellClass}>
-      <Dropdown overlay={menu}>
-        <IconButton name="ellipsis-v" aria-label="Actions" size="sm" onClick={(e) => e.stopPropagation()} />
-      </Dropdown>
-    </div>
-  );
-});
-
-function FunctionCellWithExpander({
-  row,
-  value,
-  depth,
-  hasChildren,
-  rowIndex,
-  rows,
-  onSymbolClick,
-  styles,
-  compact = false,
-  toggleRowExpanded,
-}: {
-  row: Row<CallTreeNode> & UseExpandedRowProps<CallTreeNode>;
-  value: string;
-  depth: number;
-  hasChildren: boolean;
-  rowIndex?: number;
-  rows: Array<Row<CallTreeNode>>;
-  onSymbolClick: (symbol: string) => void;
-  styles: Styles;
-  compact?: boolean;
-  toggleRowExpanded: (id: string[], value?: boolean) => void;
-}) {
-  const expandSingleChildChain = (node: CallTreeNode) => {
-    if (node.children?.length === 1) {
-      const childNode = node.children[0];
-      toggleRowExpanded([childNode.id], true);
-      if (childNode.children && childNode.children.length > 0) {
-        expandSingleChildChain(childNode);
-      }
-    }
-  };
-
-  const handleClick = () => {
-    if (hasChildren) {
-      const wasExpanded = row.isExpanded;
-      row.toggleRowExpanded();
-      if (!wasExpanded) {
-        expandSingleChildChain(row.original);
-      }
-    }
-    onSymbolClick(value);
-  };
-
-  const isLastVisibleChildAtIndex = (index: number): boolean => {
-    if (index === undefined) {
-      return false;
-    }
-
-    const currentRow = rows[index];
-    const parentId = currentRow.original.parentId;
-
-    for (let i = index + 1; i < rows.length; i++) {
-      if (rows[i].original.parentId === parentId) {
-        return false;
-      }
-      if (rows[i].original.depth <= currentRow.original.depth) {
-        break;
-      }
-    }
-    return true;
-  };
-
-  const buildTreeConnector = () => {
-    if (depth === 0) {
-      return null;
-    }
-
-    const lines: string[] = [];
-
-    const nodeIdToNode = new Map<string, CallTreeNode>();
-    rows.forEach((r) => {
-      nodeIdToNode.set(r.original.id, r.original);
-    });
-
-    const hasMoreNodesAtDepth = (targetDepth: number): boolean => {
-      if (rowIndex === undefined) {
-        return false;
-      }
-
-      for (let i = rowIndex + 1; i < rows.length; i++) {
-        const checkRow = rows[i];
-
-        if (checkRow.original.depth === targetDepth) {
-          return true;
-        }
-
-        if (checkRow.original.depth < targetDepth) {
-          break;
-        }
-      }
-      return false;
-    };
-
-    let currentNode = row.original;
-
-    while (currentNode.parentId && currentNode.depth > 0) {
-      const parent = nodeIdToNode.get(currentNode.parentId);
-      if (parent) {
-        currentNode = parent;
-      } else {
-        break;
-      }
-    }
-
-    for (let i = 0; i < depth - 1; i++) {
-      if (hasMoreNodesAtDepth(i + 1)) {
-        lines.push('│ ');
-      } else {
-        lines.push('  ');
-      }
-    }
-
-    const isLastChild = rowIndex !== undefined ? isLastVisibleChildAtIndex(rowIndex) : false;
-    lines.push(isLastChild ? '└─' : '├─');
-
-    return lines.join('');
-  };
-
-  const connector = buildTreeConnector();
-
-  return (
-    <div className={styles.functionCellContainer}>
-      {connector && <span className={styles.treeConnector}>{connector} </span>}
-      <span className={styles.functionNameWrapper}>
-        <Button fill="text" size="sm" onClick={handleClick} className={styles.functionButton}>
-          {value}
-        </Button>
-        {!compact && row.original.children && row.original.children.length > 0 && (
-          <span className={styles.nodeBadge}>
-            {row.original.children.length} {row.original.children.length === 1 ? 'child' : 'children'},{' '}
-            {row.original.subtreeSize} {row.original.subtreeSize === 1 ? 'node' : 'nodes'}
-          </span>
-        )}
-      </span>
-    </div>
-  );
-}
-
-function ColorBarCell({
-  node,
-  data,
-  colorScheme,
-  theme,
-  styles,
-  focusedNode,
-}: {
-  node: CallTreeNode;
-  data: FlameGraphDataContainer;
-  colorScheme: ColorScheme | ColorSchemeDiff;
-  theme: GrafanaTheme2;
-  styles: Styles;
-  focusedNode?: CallTreeNode;
-}) {
-  const barColor = getRowBarColor(node, data, colorScheme, theme);
-
-  let barWidth: string;
-
-  if (focusedNode) {
-    if (node.id === focusedNode.parentId) {
-      barWidth = '0%';
-    } else {
-      const relativePercent = focusedNode.total > 0 ? (node.total / focusedNode.total) * 100 : 0;
-      barWidth = `${Math.min(relativePercent, 100)}%`;
-    }
-  } else {
-    barWidth = `${Math.min(node.totalPercent, 100)}%`;
-  }
-
-  return (
-    <div className={styles.colorBarContainer}>
-      <div className={styles.colorBar} style={{ width: barWidth, backgroundColor: barColor }} />
-    </div>
-  );
-}
-
-function DiffCell({ value, theme }: { value: number | undefined; theme: GrafanaTheme2 }) {
-  if (value === undefined) {
-    return <span>-</span>;
-  }
-
-  let displayValue: string;
-  let color: string;
-
-  if (value === Infinity) {
-    displayValue = 'new';
-    color = theme.colors.success.text;
-  } else if (value === -100) {
-    displayValue = 'removed';
-    color = theme.colors.error.text;
-  } else {
-    displayValue = `${value > 0 ? '+' : ''}${value.toFixed(2)}%`;
-    color = value > 0 ? theme.colors.error.text : theme.colors.success.text;
-  }
-
-  return <span style={{ color, fontWeight: 'bold' }}>{displayValue}</span>;
-}
-
-function getStyles(theme: GrafanaTheme2) {
-  return {
-    container: css({
-      width: '100%',
-      height: '100%',
-      backgroundColor: theme.colors.background.primary,
-      display: 'flex',
-      flexDirection: 'column',
-    }),
-    scrollContainer: css({
-      '&::-webkit-scrollbar': {
-        width: '8px',
-      },
-      '&::-webkit-scrollbar-track': {
-        background: theme.colors.background.secondary,
-      },
-      '&::-webkit-scrollbar-thumb': {
-        background: theme.colors.text.disabled,
-        borderRadius: theme.shape.radius.default,
-      },
-      '&::-webkit-scrollbar-thumb:hover': {
-        background: theme.colors.text.secondary,
-      },
-    }),
-    toolbar: css({
-      display: 'flex',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      paddingTop: theme.spacing(1),
-      paddingBottom: theme.spacing(1),
-      gap: theme.spacing(1),
-      flexWrap: 'wrap',
-      borderBottom: `1px solid ${theme.colors.border.weak}`,
-    }),
-    toolbarLeft: css({
-      display: 'flex',
-      alignItems: 'center',
-      gap: theme.spacing(1),
-    }),
-    toolbarRight: css({
-      display: 'flex',
-      alignItems: 'center',
-      flexWrap: 'wrap',
-      gap: theme.spacing(1),
-    }),
-    searchContainer: css({
-      display: 'flex',
-      alignItems: 'center',
-      gap: theme.spacing(1),
-      flexWrap: 'wrap',
-    }),
-    searchNavigation: css({
-      display: 'flex',
-      alignItems: 'center',
-      gap: theme.spacing(0.5),
-      padding: `0 ${theme.spacing(1)}`,
-    }),
-    searchCounter: css({
-      fontSize: theme.typography.bodySmall.fontSize,
-      color: theme.colors.text.secondary,
-      whiteSpace: 'nowrap',
-    }),
-    searchNoResults: css({
-      fontSize: theme.typography.bodySmall.fontSize,
-      color: theme.colors.text.secondary,
-      fontStyle: 'italic',
-    }),
-    searchError: css({
-      fontSize: theme.typography.bodySmall.fontSize,
-      color: theme.colors.error.text,
-    }),
-    table: css({
-      width: '100%',
-      tableLayout: 'fixed',
-      borderCollapse: 'collapse',
-      fontSize: theme.typography.fontSize,
-      color: theme.colors.text.primary,
-    }),
-    thead: css({
-      backgroundColor: theme.colors.background.secondary,
-    }),
-    th: css({
-      padding: '4px 6px',
-      textAlign: 'left',
-      fontWeight: theme.typography.fontWeightMedium,
-      borderBottom: `1px solid ${theme.colors.border.weak}`,
-      cursor: 'pointer',
-      userSelect: 'none',
-      '&:hover': {
-        backgroundColor: theme.colors.emphasize(theme.colors.background.secondary, 0.03),
-      },
-    }),
-    tbody: css({
-      backgroundColor: theme.colors.background.primary,
-    }),
-    tr: css({
-      '&:hover': {
-        backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.03),
-      },
-    }),
-    focusedRow: css({
-      backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.08),
-      borderLeft: `3px solid ${theme.colors.primary.main}`,
-      fontWeight: theme.typography.fontWeightMedium,
-      '&:hover': {
-        backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.1),
-      },
-    }),
-    callersTargetRow: css({
-      backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.08),
-      borderLeft: `3px solid ${theme.colors.info.main}`,
-      fontWeight: theme.typography.fontWeightMedium,
-      '&:hover': {
-        backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.1),
-      },
-    }),
-    searchMatchRow: css({
-      backgroundColor: theme.colors.warning.transparent,
-      borderLeft: `3px solid ${theme.colors.warning.main}`,
-      fontWeight: theme.typography.fontWeightMedium,
-      '&:hover': {
-        backgroundColor: theme.colors.emphasize(theme.colors.warning.transparent, 0.1),
-      },
-    }),
-    td: css({
-      padding: '0px 6px',
-      borderBottom: 'none',
-      height: '20px',
-      verticalAlign: 'middle',
-      overflow: 'hidden',
-    }),
-    valueCell: css({
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'space-between',
-      gap: '8px',
-      fontVariantNumeric: 'tabular-nums',
-      height: '20px',
-    }),
-    valueNumber: css({
-      flex: '1 1 auto',
-      textAlign: 'right',
-      whiteSpace: 'nowrap',
-      minWidth: '60px',
-    }),
-    percentNumber: css({
-      flex: '0 0 60px',
-      width: '60px',
-      textAlign: 'right',
-      color: theme.colors.text.secondary,
-      whiteSpace: 'nowrap',
-    }),
-    functionCellContainer: css({
-      display: 'flex',
-      alignItems: 'center',
-      gap: '2px',
-      height: '20px',
-      lineHeight: '1',
-      overflow: 'hidden',
-      minWidth: 0,
-    }),
-    treeConnector: css({
-      color: theme.colors.text.secondary,
-      fontSize: '16px',
-      lineHeight: '1',
-      fontFamily: 'monospace',
-      whiteSpace: 'pre',
-      display: 'inline-block',
-      verticalAlign: 'middle',
-      flexShrink: 0,
-    }),
-    functionNameWrapper: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      overflow: 'hidden',
-      minWidth: 0,
-    }),
-    functionButton: css({
-      padding: 0,
-      fontSize: theme.typography.fontSize,
-      textAlign: 'left',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-      minWidth: 0,
-      flexShrink: 1,
-    }),
-    nodeBadge: css({
-      marginLeft: theme.spacing(0.5),
-      fontSize: theme.typography.bodySmall.fontSize,
-      color: theme.colors.text.secondary,
-      whiteSpace: 'nowrap',
-      flexShrink: 0,
-    }),
-    sortIndicator: css({
-      marginLeft: '4px',
-      fontSize: '10px',
-    }),
-    colorBarContainer: css({
-      width: '100%',
-      height: '20px',
-      display: 'flex',
-      alignItems: 'center',
-    }),
-    colorBar: css({
-      height: '16px',
-      minWidth: '2px',
-      borderRadius: theme.shape.radius.default,
-    }),
-    actionsCell: css({
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      height: '20px',
-    }),
-    actionsColumnCell: css({
-      backgroundColor: theme.colors.background.secondary,
-      '&:hover': {
-        backgroundColor: theme.colors.background.secondary,
-      },
-    }),
-    valueColumnCell: css({
-      overflow: 'visible',
-      textAlign: 'right',
-    }),
-    modePill: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      background: theme.colors.background.secondary,
-      borderRadius: theme.shape.borderRadius(8),
-      padding: theme.spacing(0.5, 1),
-      fontSize: theme.typography.bodySmall.fontSize,
-      fontWeight: theme.typography.fontWeightMedium,
-      lineHeight: theme.typography.bodySmall.lineHeight,
-      color: theme.colors.text.secondary,
-    }),
-    modePillLabel: css({
-      maxWidth: '200px',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-      marginLeft: theme.spacing(0.5),
-    }),
-    modePillCloseButton: css({
-      verticalAlign: 'text-bottom',
-      margin: theme.spacing(0, 0.5),
-    }),
-    focusedItem: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      background: theme.colors.background.secondary,
-      borderRadius: theme.shape.radius.default,
-      padding: theme.spacing(0.5, 1),
-      fontSize: theme.typography.bodySmall.fontSize,
-      fontWeight: theme.typography.fontWeightMedium,
-      lineHeight: theme.typography.bodySmall.lineHeight,
-      color: theme.colors.text.secondary,
-    }),
-    focusedItemLabel: css({
-      maxWidth: '200px',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-      marginLeft: theme.spacing(0.5),
-    }),
-    callersItem: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      background: theme.colors.background.secondary,
-      borderRadius: theme.shape.radius.default,
-      padding: theme.spacing(0.5, 1),
-      fontSize: theme.typography.bodySmall.fontSize,
-      fontWeight: theme.typography.fontWeightMedium,
-      lineHeight: theme.typography.bodySmall.lineHeight,
-      color: theme.colors.text.secondary,
-    }),
-    callersItemLabel: css({
-      maxWidth: '200px',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-      marginLeft: theme.spacing(0.5),
-    }),
-  };
-}
 
 export default FlameGraphCallTreeContainer;
