+++ packages/grafana-flamegraph/src/CallTree/styles.ts
@@ -1,0 +1,291 @@
+import { css } from '@emotion/css';
+
+import { GrafanaTheme2 } from '@grafana/data';
+
+export type Styles = ReturnType<typeof getStyles>;
+
+export function getStyles(theme: GrafanaTheme2) {
+  return {
+    container: css({
+      width: '100%',
+      height: '100%',
+      backgroundColor: theme.colors.background.primary,
+      display: 'flex',
+      flexDirection: 'column',
+    }),
+    scrollContainer: css({
+      '&::-webkit-scrollbar': {
+        width: '8px',
+      },
+      '&::-webkit-scrollbar-track': {
+        background: theme.colors.background.secondary,
+      },
+      '&::-webkit-scrollbar-thumb': {
+        background: theme.colors.text.disabled,
+        borderRadius: theme.shape.radius.default,
+      },
+      '&::-webkit-scrollbar-thumb:hover': {
+        background: theme.colors.text.secondary,
+      },
+    }),
+    toolbar: css({
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      paddingTop: theme.spacing(1),
+      paddingBottom: theme.spacing(1),
+      gap: theme.spacing(1),
+      flexWrap: 'wrap',
+      borderBottom: `1px solid ${theme.colors.border.weak}`,
+    }),
+    toolbarLeft: css({
+      display: 'flex',
+      alignItems: 'center',
+      gap: theme.spacing(1),
+    }),
+    toolbarRight: css({
+      display: 'flex',
+      alignItems: 'center',
+      flexWrap: 'wrap',
+      gap: theme.spacing(1),
+    }),
+    searchContainer: css({
+      display: 'flex',
+      alignItems: 'center',
+      gap: theme.spacing(1),
+      flexWrap: 'wrap',
+    }),
+    searchNavigation: css({
+      display: 'flex',
+      alignItems: 'center',
+      gap: theme.spacing(0.5),
+      padding: `0 ${theme.spacing(1)}`,
+    }),
+    searchCounter: css({
+      fontSize: theme.typography.bodySmall.fontSize,
+      color: theme.colors.text.secondary,
+      whiteSpace: 'nowrap',
+    }),
+    searchNoResults: css({
+      fontSize: theme.typography.bodySmall.fontSize,
+      color: theme.colors.text.secondary,
+      fontStyle: 'italic',
+    }),
+    searchError: css({
+      fontSize: theme.typography.bodySmall.fontSize,
+      color: theme.colors.error.text,
+    }),
+    table: css({
+      width: '100%',
+      tableLayout: 'fixed',
+      borderCollapse: 'collapse',
+      fontSize: theme.typography.fontSize,
+      color: theme.colors.text.primary,
+    }),
+    thead: css({
+      backgroundColor: theme.colors.background.secondary,
+    }),
+    th: css({
+      padding: '4px 6px',
+      textAlign: 'left',
+      fontWeight: theme.typography.fontWeightMedium,
+      borderBottom: `1px solid ${theme.colors.border.weak}`,
+      cursor: 'pointer',
+      userSelect: 'none',
+      '&:hover': {
+        backgroundColor: theme.colors.emphasize(theme.colors.background.secondary, 0.03),
+      },
+    }),
+    tbody: css({
+      backgroundColor: theme.colors.background.primary,
+    }),
+    tr: css({
+      '&:hover': {
+        backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.03),
+      },
+    }),
+    focusedRow: css({
+      backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.08),
+      borderLeft: `3px solid ${theme.colors.primary.main}`,
+      fontWeight: theme.typography.fontWeightMedium,
+      '&:hover': {
+        backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.1),
+      },
+    }),
+    callersTargetRow: css({
+      backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.08),
+      borderLeft: `3px solid ${theme.colors.info.main}`,
+      fontWeight: theme.typography.fontWeightMedium,
+      '&:hover': {
+        backgroundColor: theme.colors.emphasize(theme.colors.background.primary, 0.1),
+      },
+    }),
+    searchMatchRow: css({
+      backgroundColor: theme.colors.warning.transparent,
+      borderLeft: `3px solid ${theme.colors.warning.main}`,
+      fontWeight: theme.typography.fontWeightMedium,
+      '&:hover': {
+        backgroundColor: theme.colors.emphasize(theme.colors.warning.transparent, 0.1),
+      },
+    }),
+    td: css({
+      padding: '0px 6px',
+      borderBottom: 'none',
+      height: '20px',
+      verticalAlign: 'middle',
+      overflow: 'hidden',
+    }),
+    valueCell: css({
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+      gap: '8px',
+      fontVariantNumeric: 'tabular-nums',
+      height: '20px',
+    }),
+    valueNumber: css({
+      flex: '1 1 auto',
+      textAlign: 'right',
+      whiteSpace: 'nowrap',
+      minWidth: '60px',
+    }),
+    percentNumber: css({
+      flex: '0 0 60px',
+      width: '60px',
+      textAlign: 'right',
+      color: theme.colors.text.secondary,
+      whiteSpace: 'nowrap',
+    }),
+    functionCellContainer: css({
+      display: 'flex',
+      alignItems: 'center',
+      gap: '2px',
+      height: '20px',
+      lineHeight: '1',
+      overflow: 'hidden',
+      minWidth: 0,
+    }),
+    treeConnector: css({
+      color: theme.colors.text.secondary,
+      fontSize: '16px',
+      lineHeight: '1',
+      fontFamily: 'monospace',
+      whiteSpace: 'pre',
+      display: 'inline-block',
+      verticalAlign: 'middle',
+      flexShrink: 0,
+    }),
+    functionNameWrapper: css({
+      display: 'inline-flex',
+      alignItems: 'center',
+      overflow: 'hidden',
+      minWidth: 0,
+    }),
+    functionButton: css({
+      padding: 0,
+      fontSize: theme.typography.fontSize,
+      textAlign: 'left',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+      minWidth: 0,
+      flexShrink: 1,
+    }),
+    nodeBadge: css({
+      marginLeft: theme.spacing(0.5),
+      fontSize: theme.typography.bodySmall.fontSize,
+      color: theme.colors.text.secondary,
+      whiteSpace: 'nowrap',
+      flexShrink: 0,
+    }),
+    sortIndicator: css({
+      marginLeft: '4px',
+      fontSize: '10px',
+    }),
+    colorBarContainer: css({
+      width: '100%',
+      height: '20px',
+      display: 'flex',
+      alignItems: 'center',
+    }),
+    colorBar: css({
+      height: '16px',
+      minWidth: '2px',
+      borderRadius: theme.shape.radius.default,
+    }),
+    actionsCell: css({
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      height: '20px',
+    }),
+    actionsColumnCell: css({
+      backgroundColor: theme.colors.background.secondary,
+      '&:hover': {
+        backgroundColor: theme.colors.background.secondary,
+      },
+    }),
+    valueColumnCell: css({
+      overflow: 'visible',
+      textAlign: 'right',
+    }),
+    modePill: css({
+      display: 'inline-flex',
+      alignItems: 'center',
+      background: theme.colors.background.secondary,
+      borderRadius: theme.shape.borderRadius(8),
+      padding: theme.spacing(0.5, 1),
+      fontSize: theme.typography.bodySmall.fontSize,
+      fontWeight: theme.typography.fontWeightMedium,
+      lineHeight: theme.typography.bodySmall.lineHeight,
+      color: theme.colors.text.secondary,
+    }),
+    modePillLabel: css({
+      maxWidth: '200px',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+      marginLeft: theme.spacing(0.5),
+    }),
+    modePillCloseButton: css({
+      verticalAlign: 'text-bottom',
+      margin: theme.spacing(0, 0.5),
+    }),
+    focusedItem: css({
+      display: 'inline-flex',
+      alignItems: 'center',
+      background: theme.colors.background.secondary,
+      borderRadius: theme.shape.radius.default,
+      padding: theme.spacing(0.5, 1),
+      fontSize: theme.typography.bodySmall.fontSize,
+      fontWeight: theme.typography.fontWeightMedium,
+      lineHeight: theme.typography.bodySmall.lineHeight,
+      color: theme.colors.text.secondary,
+    }),
+    focusedItemLabel: css({
+      maxWidth: '200px',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+      marginLeft: theme.spacing(0.5),
+    }),
+    callersItem: css({
+      display: 'inline-flex',
+      alignItems: 'center',
+      background: theme.colors.background.secondary,
+      borderRadius: theme.shape.radius.default,
+      padding: theme.spacing(0.5, 1),
+      fontSize: theme.typography.bodySmall.fontSize,
+      fontWeight: theme.typography.fontWeightMedium,
+      lineHeight: theme.typography.bodySmall.lineHeight,
+      color: theme.colors.text.secondary,
+    }),
+    callersItemLabel: css({
+      maxWidth: '200px',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+      marginLeft: theme.spacing(0.5),
+    }),
+  };
+}
